+++ src/c/core/session/stream_id.c
@@ -20,10 +20,10 @@ mrStreamId mr_stream_id(uint8_t index, mrStreamType type, mrStreamDirection dire
             stream_id.raw = index;
             break;
         case MR_BEST_EFFORT_STREAM:
-            stream_id.raw = index + BEST_EFFORT_STREAM_THRESHOLD;
+            stream_id.raw = (uint8_t)(index + BEST_EFFORT_STREAM_THRESHOLD);
             break;
         case MR_RELIABLE_STREAM:
-            stream_id.raw = index + RELIABLE_STREAM_THRESHOLD;
+            stream_id.raw = (uint8_t)(index + RELIABLE_STREAM_THRESHOLD);
             break;
     }
 
@@ -43,12 +43,12 @@ mrStreamId mr_stream_id_from_raw(uint8_t stream_id_raw, mrStreamDirection direct
     }
     else if(RELIABLE_STREAM_THRESHOLD > stream_id_raw)
     {
-        stream_id.index = stream_id_raw - BEST_EFFORT_STREAM_THRESHOLD;
+        stream_id.index = (uint8_t)(stream_id_raw - BEST_EFFORT_STREAM_THRESHOLD);
         stream_id.type = MR_BEST_EFFORT_STREAM;
     }
     else
     {
-        stream_id.index = stream_id_raw - RELIABLE_STREAM_THRESHOLD;
+        stream_id.index = (uint8_t)(stream_id_raw - RELIABLE_STREAM_THRESHOLD);
         stream_id.type = MR_RELIABLE_STREAM;
     }
